# backend/openapi/openapi.json
{"openapi":"3.0.0","paths":{"/mashup/auth":{"post":{"operationId":"MashupController_checkAuth","parameters":[],"responses":{"201":{"description":""},"403":{"description":""}},"tags":["mashup"],"security":[{"bearer":[]}]}},"/mashup/artist":{"get":{"operationId":"MashupController_searchArtist","parameters":[{"name":"name","required":true,"in":"query","description":"The name of the artist to search","example":"Kanye West","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchArtistResponseSchema"}}}},"403":{"description":""}},"tags":["mashup"],"security":[{"bearer":[]}]}},"/mashup/playlists":{"post":{"operationId":"MashupController_generatePlaylist","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmitPlaylistSchema"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeneratePlaylistResponseSchema"}}}},"403":{"description":""}},"tags":["mashup"],"security":[{"bearer":[]}]},"get":{"operationId":"MashupController_getPlaylists","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetPlaylistResponseSchema"}}}}},"403":{"description":""}},"tags":["mashup"],"security":[{"bearer":[]}]}},"/mashup/playlists/{playlist}":{"put":{"operationId":"MashupController_updatePlaylist","parameters":[{"name":"playlist","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePlaylistSchema"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeneratePlaylistResponseSchema"}}}},"403":{"description":""}},"tags":["mashup"],"security":[{"bearer":[]}]},"get":{"operationId":"MashupController_getPlaylist","parameters":[{"name":"playlist","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetPlaylistResponseSchema"}}}},"403":{"description":""}},"tags":["mashup"],"security":[{"bearer":[]}]}},"/mashup/playlists/{playlist}/refresh":{"post":{"operationId":"MashupController_refreshPlaylist","parameters":[{"name":"playlist","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""},"403":{"description":""}},"tags":["mashup"],"security":[{"bearer":[]}]}},"/mashup/playlists/{playlist}/active":{"post":{"operationId":"MashupController_setPlaylistActive","parameters":[{"name":"playlist","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""},"403":{"description":""}},"tags":["mashup"],"security":[{"bearer":[]}]}},"/mashup/playlists/{playlist}/inactive":{"post":{"operationId":"MashupController_setPlaylistInactive","parameters":[{"name":"playlist","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""},"403":{"description":""}},"tags":["mashup"],"security":[{"bearer":[]}]}},"/mashup/deleteUser":{"post":{"operationId":"MashupController_deleteUser","parameters":[],"responses":{"204":{"description":""},"403":{"description":""}},"tags":["mashup"],"security":[{"bearer":[]}]}}},"info":{"title":"Spotify apps NestJS backend","description":"The API for the Spotify apps","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"ArtistResponseFull":{"type":"object","properties":{"images":{"description":"Array of URLs to artist images","type":"array","items":{"type":"string"}},"number":{"type":"number","description":"Number of songs to select from the artist","example":25},"playlist":{"type":"string","description":"The ID of the 'This is XYZ' playlist","example":"37i9dQZF1DZ06evO3nMr04"},"id":{"type":"string","description":"The ID of the artist","example":"5K4W6rqBFWDnAN6FQUkS6x"},"name":{"type":"string","description":"The name of the artist","example":"Kanye West"}},"required":["images","number","playlist","id","name"]},"SearchArtistResponseSchema":{"type":"object","properties":{"query":{"type":"string","description":"The search query","example":"Ye"},"errorReason":{"type":"string","nullable":true,"description":"The reason in case a search fails"},"artist":{"nullable":true,"description":"The artist that is returned","allOf":[{"$ref":"#/components/schemas/ArtistResponseFull"}]}},"required":["query","errorReason","artist"]},"ArtistFull":{"type":"object","properties":{"number":{"type":"number","description":"Number of songs to select from the artist","example":25},"playlist":{"type":"string","description":"The ID of the 'This is XYZ' playlist","example":"37i9dQZF1DZ06evO3nMr04"},"id":{"type":"string","description":"The ID of the artist","example":"5K4W6rqBFWDnAN6FQUkS6x"},"name":{"type":"string","description":"The name of the artist","example":"Kanye West"}},"required":["number","playlist","id","name"]},"SubmitPlaylistSchema":{"type":"object","properties":{"parts":{"description":"Contains IDs of 'This is XYZ' playlists and their artists IDs and names","type":"array","items":{"$ref":"#/components/schemas/ArtistFull"}},"active":{"type":"boolean","description":"Defines if playlist refreshing is set to active or not","example":true},"frequency":{"type":"number","description":"Refresh playlist every X days","example":7}},"required":["parts","active","frequency"]},"GeneratePlaylistResponseSchema":{"type":"object","properties":{"id":{"type":"string","description":"The ID of the generated playlist","example":"7EHw08q3YBIvG9iGI7Q9Ob"}},"required":["id"]},"UpdatePlaylistSchema":{"type":"object","properties":{"parts":{"description":"Contains IDs of 'This is XYZ' playlists and their artists IDs and names","type":"array","items":{"$ref":"#/components/schemas/ArtistFull"}},"active":{"type":"boolean","description":"Defines if playlist refreshing is set to active or not","example":true},"frequency":{"type":"number","description":"Refresh playlist every X days","example":7},"updateTitle":{"type":"boolean","description":"Defines if the title of the playlist should be updated","example":true},"updateDescription":{"type":"boolean","description":"Defines if the description of the playlist should be updated","example":true},"updateSongs":{"type":"boolean","description":"Defines if the songs of the playlist should be updated","example":true},"updateMetadata":{"type":"boolean","description":"Defines if the metadata of the playlist should be updated","example":true}},"required":["parts","active","frequency","updateTitle","updateDescription","updateSongs","updateMetadata"]},"PlaylistsResponse":{"type":"object","properties":{"name":{"type":"string","description":"The name of the playlist","example":"These are Labrinth, Sia and Diplo"},"description":{"type":"string","description":"The description of the playlist","example":"This playlist was auto-generated! Artists are Diplo, Labrinth and Sia."},"images":{"description":"Array of URLs to playlist images","type":"array","items":{"type":"string"}}},"required":["name","description","images"]},"MetadataResponse":{"type":"object","properties":{"updated":{"type":"number","description":"Timestamp of last update","example":1661433338296},"created":{"type":"number","description":"Timestamp of creation","example":1661433338296},"active":{"type":"boolean","description":"Defines if playlist refreshing is set to active or not","example":true},"frequency":{"type":"number","description":"Refresh playlist every X days","example":7}},"required":["updated","created","active","frequency"]},"GetPlaylistResponseSchema":{"type":"object","properties":{"id":{"type":"string","description":"The ID of the playlist","example":"7EHw08q3YBIvG9iGI7Q9Ob"},"playlist":{"$ref":"#/components/schemas/PlaylistsResponse"},"artists":{"type":"array","items":{"$ref":"#/components/schemas/ArtistResponseFull"}},"metadata":{"$ref":"#/components/schemas/MetadataResponse"}},"required":["id","playlist","artists","metadata"]}}}}